feat: improve landing page
