+++ src/lib/components/common/Footer.svelte
@@ -98,14 +98,6 @@
 							{m['footer.github']()}
 						</a>
 					</li>
-					<li>
-						<a
-							href="/learn-more"
-							class="text-muted-foreground transition-colors hover:text-foreground"
-						>
-							{m['footer.learnMore']()}
-						</a>
-					</li>
 				</ul>
 			</div>
